Extract AIConfigList
